+++ CHANGELOG.md
@@ -8,12 +8,35 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Added info to contributors (CONTRIBUTING.md).
+  ([62949a1](https://github.com/scriptotek/php-marc/commit/62949a1b2e1c309e3bf8bb58f9f8f138c0398d46))
+- Added initialization from SimpleXMLElement object through the new methods
+  `Collection:fromSimpleXMLElement($obj)` and `Record:fromSimpleXMLElement($obj)`.
+
+### Fixed
+
+- Improved documentation and support for IDE code analysis.
+  ([#15](https://github.com/scriptotek/php-marc/issues/15)
+  by [@rudolfbyker](https://github.com/rudolfbyker))
+
+## [2.0.2] - 2019-09-13
+
+### Added
+
 - Added new method `Field::asLineMarc()` to return a line mode Marc string
   representation of the field.
+  ([ba20a6d](https://github.com/scriptotek/php-marc/commit/ba20a6deadc9402bb65807cd63e33797d2893dea))
 
 ### Fixed
 
 - Fixed the `Subject::getParts()` method.
+  ([1fe8408](https://github.com/scriptotek/php-marc/commit/1fe8408e49c6c3afba9ec379b441c82f64ce0336))
+- Added additional subject subfield codes that were missing.
+  ([7908616](https://github.com/scriptotek/php-marc/commit/79086165dfce9b9d2f490d38e9f50f70fef5641f))
+- Added 852 $i and $j to `Location.callCode`.
+  ([cba1508](https://github.com/scriptotek/php-marc/commit/cba15083422bb2ac812b6b355341feab2cff308a))
+- Fixed the string representation of the `Location` class.
+  ([74652a3](https://github.com/scriptotek/php-marc/commit/74652a3bf4cc3e9fe3c916057a0a9bd47419f601))
 
 ## [2.0.1] - 2019-01-09
 
@@ -90,7 +113,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Record::fromString` now throws a `RecordNotFound` exception rather than an `ErrorException` exception if no record was found.
 - `Record::getType` now throws a `UnknownRecordType` exception rather than an `ErrorException`.
 
-[Unreleased]: https://github.com/scriptotek/php-marc/compare/v2.0.1...HEAD
+[Unreleased]: https://github.com/scriptotek/php-marc/compare/v2.0.2...HEAD
+[2.0.2]: https://github.com/scriptotek/php-marc/compare/v2.0.1...v2.0.2
 [2.0.1]: https://github.com/scriptotek/php-marc/compare/v2.0.0...v2.0.1
 [2.0.0]: https://github.com/scriptotek/php-marc/compare/v1.0.1...v2.0.0
 [1.0.1]: https://github.com/scriptotek/php-marc/compare/v1.0.0...v1.0.1